Initial working prototype. This works as data entry into a local CSV file (included) with a command line interface.
original description:

This is a quick console data entry program for the Five Element Nutrition theory of Traditional Chinese Medicine (TCM). It is designed to input foods, herbal medicines, or anything consumable and classify it quickly according to organ or element, then associated temperature(s) (qualities of the food). A GUI could easily be added however console input is very simple, so I am starting with this. This application is designed to show a quick way to enter data for a specific purpose. This reads and writes to a CSV file of foods called 'foodList.csv' if it is absent it will be created. In the future this could link to an online spreadsheet allowing multiple contributers to the dataset, as well as  keeping integrity of the data. For now the spreadsheet is stored locally. If you are interested in more of the back-story behind this application, or are wish to get involved please let me know.